pyproject.toml
[project]
name = "dte-core-engine"
version = "0.1.0"
description = "Motor DTE para facturación electrónica SII Chile"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "zeep>=4.2.1",
    "lxml>=5.1.0",
    "signxml>=3.2.0",
    "cryptography>=42.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "httpx>=0.27.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
